# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/all": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter users by role",
            "schema": {
              "example": "therapist",
              "type": "string",
              "enum": [
                "admin",
                "patient",
                "therapist"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "hashcklsachlka"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (or of specific roles)",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "hashcklsachlka"
                }
              }
            }
          }
        },
        "summary": "User Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/patient/register": {
      "post": {
        "operationId": "AuthController_registerPatient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new patient",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/therapist/register": {
      "post": {
        "operationId": "AuthController_registerTherapist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTherapistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new therapist",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/register": {
      "post": {
        "operationId": "AuthController_registerAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new admin (Can only be used by existing admins)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/request-password-reset": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password request email has been sent"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Request password reset email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password has been reset successfully"
          },
          "400": {
            "description": "Token is invalid or has expired"
          }
        },
        "summary": "Reset password using token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/therapy-types": {
      "get": {
        "operationId": "TherapyTypeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of therapy types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapyTypeResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all therapy types",
        "tags": [
          "TherapyType"
        ]
      },
      "post": {
        "operationId": "TherapyTypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTherapyTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Therapy type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapyTypeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new therapy type (Admin only route)",
        "tags": [
          "TherapyType"
        ]
      }
    },
    "/therapy-types/{id}": {
      "get": {
        "operationId": "TherapyTypeController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Therapy type ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Therapy type found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapyTypeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a therapy type by ID",
        "tags": [
          "TherapyType"
        ]
      },
      "put": {
        "operationId": "TherapyTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Therapy type ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTherapyTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Therapy type updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapyTypeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a therapy type (Admin only route)",
        "tags": [
          "TherapyType"
        ]
      },
      "delete": {
        "operationId": "TherapyTypeController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Therapy type ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Therapy type has been deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a therapy type (Admin only route)",
        "tags": [
          "TherapyType"
        ]
      }
    },
...
  },
  "info": {
    "title": "Therapist Booking System API",
    "description": "API documentation for the therapist booking system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginPatientDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterPatientDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password",
            "minLength": 6
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the user"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterTherapistDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password",
            "minLength": 6
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the user"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "RegisterAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email of the user"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password",
            "minLength": 6
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name of the user"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com",
            "description": "Email of the user"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "xgsauigaskuhkas3629)",
            "description": "A string token"
          },
          "newPassword": {
            "type": "string",
            "example": "password",
            "description": "New Password of the user"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "TherapyTypeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Unique identifier for the therapy type"
          },
          "name": {
            "type": "string",
            "example": "Individual Therapy",
            "description": "The name of the therapy type"
          },
          "description": {
            "type": "string",
            "example": "One-on-one therapy session",
            "description": "A short description of the therapy type"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-01T12:00:00Z",
            "description": "Timestamp when the therapy type was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-02T12:00:00Z",
            "description": "Timestamp when the therapy type was last updated"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTherapyTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Individual Therapy",
            "description": "The name of the therapy type"
          },
          "description": {
            "type": "string",
            "example": "One-on-one therapy session",
            "description": "A short description of the therapy type"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTherapyTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Individual Therapy",
            "description": "The name of the therapy type"
          },
          "description": {
            "type": "string",
            "example": "One-on-one therapy session",
            "description": "A short description of the therapy type"
          }
        }
      },
...
    }
  }
}